chore: allow switch user & tag on runtime
